fix: keep esmfold2 confidence bfloat16 opt-in, with the numbers
Reverts the default flip in f46c061. GPU rows 1111/1112 measured the
option against the released default: 1,003 tokens wall 155.16 -> 160.92 s
with peak 14,733.3 -> 14,733.3 MiB, 2,096 tokens wall 450.88 -> 450.05 s
with peak 46,041.8 -> 46,042.3 MiB. The arm fires -- the 2,096-token peak
moves 0.5 MiB -- and buys nothing, because this port's peak is a
folding-trunk arena the confidence head has no term in. Against zero gain
the divergence from pinned native is the whole ledger, so float32 stays.

Kept from f46c061, because none of it depends on the default:

- The upstream reading, now written down in the field docstring, heads.py,
  cli.md and the CHANGELOG. ConfidenceHead.forward runs outside every
  autocast region (modeling_esmfold2.py:172-221, called at :1061, after
  the model-level bf16 context opened at :936 closes at :1030); only its
  own folding_trunk is wrapped, at :223, taking a float32 pair and
  returning pair.add_(pair_delta.float()). So the float32 re-embedding is
  deliberate rather than an island nobody chose.
- The measured rows, in cli.md and the CHANGELOG, and the statement that
  the option is for combination arms and is not recommended alone.
- validate_native_options reads the default from _FIXED_COMPILE_DEFAULTS
  instead of repeating the literal.
- The two-layer drift test pinning _FIXED_COMPILE_DEFAULTS against
  ModelSettings, and the default test now resolving the width through the
  settings and asserting realised operand dtypes rather than the string.
- heads.py:305 cited trunk.py:125 and :161 for the gate and SwiGLU
  upcasts; those lines are now an einsum equation and a parameter name.
  Corrected to :120 and :164.

+++ src/foldjax/backends/esmfold2.py
@@ -82,11 +82,10 @@ _FIXED_COMPILE_DEFAULTS = {
     # The unfused transition is what every released number describes, and the
     # value is independent of the checkpoint: no `config.json` key reaches it.
     "glu_backend": "xla",
-    # The port's own default since 2026-09-11, and no `config.json` key
-    # reaches it, so naming it explicitly must not fork a namespace. Unlike
-    # the entries above, it is *not* what the recorded numbers describe:
-    # `float32` is upstream's realised width and keeps its own namespace.
-    "confidence_dtype": "bfloat16",
+    # A float32 confidence re-embedding is what every released number
+    # describes, it is upstream's own width, and no `config.json` key reaches
+    # it either.
+    "confidence_dtype": "float32",
 }
 
 #: The values `glu_backend` accepts, spelled here rather than imported.
@@ -429,10 +428,9 @@ class ESMFold2Backend(ManagedCcdMemory, Backend):
             # rather than a neutral knob: no other port spells a GLU choice
             # the neutral vocabulary could rename, and Boltz-2's is native too.
             "glu_backend",
-            # The confidence re-embedding's width, bfloat16 by default.
-            # Native for the same reason `trunk_dtype` is: it names a region
-            # of this port's own arrangement, and the neutral vocabulary has
-            # no word for it.
+            # Opt-in bfloat16 confidence re-embedding. Native for the same
+            # reason `trunk_dtype` is: it names a region of this port's own
+            # arrangement, and the neutral vocabulary has no word for it.
             "confidence_dtype",
             "no_language_model",
             # Spelled through rather than renamed: the port's own settings
@@ -480,9 +478,8 @@ class ESMFold2Backend(ManagedCcdMemory, Backend):
         # traced against, so a fused run must not be answered out of the
         # default executable's cache entry.
         "glu_backend",
-        # Traced into the confidence head's re-embedding, so a float32
-        # opt-out must not be answered out of the narrowed executable's
-        # cache entry, or the reverse.
+        # Traced into the confidence head's re-embedding, so a narrowed run
+        # must not be answered out of the float32 executable's cache entry.
         "confidence_dtype",
     )
 
@@ -988,7 +985,7 @@ class ESMFold2Backend(ManagedCcdMemory, Backend):
         if "glu_backend" in options:
             overrides["glu_backend"] = _checked_glu_backend(options.pop("glu_backend"))
         # Absent means unasked, for the same reason: the port's own default is
-        # already `bfloat16`.
+        # already `float32`.
         if "confidence_dtype" in options:
             overrides["confidence_dtype"] = _checked_confidence_dtype(
                 options.pop("confidence_dtype")
